wanpipe-3.5.24: Fix compilation on linux-3.1.
Linux versions prior to 3.1 defined HAVE_NET_DEVICE_OPS in netdevice.h
and wanpipe relies on this to (and not the kernel version) to deceide which
set of macros to use to access netdev ops.

The new kernel dropped this precompiler constant and wanpipe falls back to
the old (pre-2.6.3X) version of macros which rely on all the netdev op function
pointer to still be in struct net_device.

Something along the lines of:

in patches/kdrivers/include/wanpipe_{kernel,defines}.h
fixes this problem. (The one in wanpipe_defines.h additionally trying
to avoid a possible redefinition warning (but maybe that's not needed;
too lazy to check)).

Signed-off-by: Stefan Knoblich <s.knoblich@axsentis.de>